[flask] create end point for /items/ GET method
Create a script that starts a flask server on a specific Host and
port. (set on config.py file)
Add an end point method that pass json list of items in a specific
time interval for a specific subreddit.
Is connecting to our mongo database and extracts data that current
exist there. (no additional requests to reddit are made over here)